move rpc to service, change to load from config
